+++ core/background.lua
@@ -1,9 +1,0 @@
-local Background = {}
-
-function Background:new(sprite)
-    local obj = {
-        sprite = sprite
-    }
-
-    return 
-end