+++ entities/grenade.lua
@@ -1,7 +1,0 @@
-local Entity = require('entity')
-
-local Grenade = {}
-Grenade.__index = Grenade
-setmetatable(Grenade, Entity)
-
-return Grenade